Goal saving endpoints and Goal+Objective Consolidation

+++ server/journal-app-server/journal-app-server/src/test/java/com/northpole/journalappserver/service/GoalTrackerServiceTest.java
@@ -61,7 +61,7 @@ public class GoalTrackerServiceTest {
                 Progress.builder()
                         .recKey("Goals Created")
                         .compareType(">=")
-                        .targetValue(1)
+                        .targetValue((double)1)
                         .build()
         );
